Remove circuit from Exporter internal state
The exporter is designed to manage state in order to perform several
exports without needing to reset the options, or re-do preprocessing
steps (or similar).  Taking a circuit as an initialisation argument
prevented the exporter from being called nicely in a loop with these
properties, which isn't the intended interface.

This change makes the interface more consistent with existing Python
exporters, such as the default JSON exporter.