Move to visitor pattern for AST printer
This fully separates out the "AST generation" step from the "conversion
to QASM" step, by moving the QASM definitions of each AST object into a
single AST visitor.  This is in anticipation of swapping the QASM 3 AST
from one defined by Terra to a reference implementation defined by the
OpenQASM team.

As a consequence, the `_qasm3_definition` attribute given to `CXGate` as
a test is changed to a method returning an AST node.  In the future, it
may be desirable to allow this function to return a string, and have
that parsed into a QASM AST node for use in the generator.  This may
become possible once the AST is removed from Terra in favour of the open
reference AST package currently in development.  Similarly, the
pretty-printer visitor may be removed from Terra once that split is
made, simplifying the code we have to take care of on our side.